refactor: Consolidate artist editor into a shared shell component
This refactors the artist creation and editing pages into a single `ArtistEditorShell` component.
The common UI and form logic are now centralized in this shell. Mode-specific data fetching, validation, and saving are handled by new `useArtistCreateFlow` and `useArtistEditFlow` composables.

This change significantly reduces code duplication, improves maintainability, and provides a consistent artist management experience.
