+++ app/pages/artist/edit/[id].vue
@@ -1,361 +1,50 @@
 <script setup lang="ts">
-	import { CalendarDate } from '@internationalized/date'
 	import { storeToRefs } from 'pinia'
 	import { useUserStore } from '~/stores/user'
 
-	import type { Artist } from '~/types'
-	import type { CompanyMenuItem } from '~/composables/useArtistEditorForm'
-	import { useMutationTimeout } from '~/composables/useMutationTimeout'
-
-	// Extended type for an artist with groups and members
-	type ArtistWithRelations = Artist & {
-		groups?: Artist[]
-		members?: Artist[]
-		releases?: unknown[]
-	}
-
 	const route = useRoute()
-	const router = useRouter()
-	const toast = useToast()
+	const title = ref('Edit Artist Page')
+	const description = ref('Edit Artist Page')
 	const userStore = useUserStore()
 	const { isAdminStore } = storeToRefs(userStore)
 
-	const { getFullArtistById, updateArtist, getSocialAndPlatformLinksByArtistId } =
-		useSupabaseArtist()
-	const { getAllMusicStyles } = useSupabaseMusicStyles()
-	const { getAllGeneralTags } = useSupabaseGeneralTags()
-	const { getAllNationalities } = useSupabaseNationalities()
-	const { runMutation } = useMutationTimeout()
-
 	const {
-		stylesList,
-		tagsList,
-		nationalitiesList,
-		artistStyles,
-		artistTags,
-		artistNationalities,
-		artistCompanies,
-		artistGroups,
-		artistMembers,
-		groupSearchTerm,
-		memberSearchTerm,
-		companySearchTerm,
-		isSearchingGroups,
-		isSearchingMembers,
-		isSearchingCompanies,
-		isCompanyModalOpen,
-		platformLinkManager,
-		socialLinkManager,
-		artistPlatformList,
-		artistSocialList,
-		genderLabels,
-		artistTypeLabels,
-		genderOptions,
-		artistTypeOptions,
-		careerOptions,
-		relationshipTypes,
-		stylesForMenu,
-		tagsForMenu,
-		nationalitiesForMenu,
-		companiesForMenu,
-		groupsForMenu,
-		membersForMenu,
-		birthdayToDate,
-		debutDateToDate,
-		birthdayInputValue,
-		debutDateInputValue,
-		formatDisplayDate,
-		buildArtistRefs,
-		buildCompanyRelationsPayload,
-		onGroupSearchTermChange,
-		onMemberSearchTermChange,
-		onCompanySearchTermChange,
-		addCompanyRelation,
-		removeCompanyRelation,
-		updateCompanyInRelation,
-		handleCompanyUpdated,
-		applyOptions,
-		applyArtistSelections,
-	} = useArtistEditorForm()
+		original,
+		model,
+		isBootstrapping,
+		bootstrapError,
+		isSaving,
+		heroImageSrc,
+		canSave,
+		bootstrap,
+		refreshNationalities,
+		save,
+	} = useArtistEditFlow(String(route.params.id ?? ''))
 
-	const title = ref('Edit Artist Page')
-	const description = ref('Edit Artist Page')
-
-	const isUploadingEdit = ref<boolean>(false)
-	const isBootstrapping = ref(true)
-	const bootstrapError = ref<string | null>(null)
-
-	const artist = ref<ArtistWithRelations | null>(null)
-
-	const artistToEdit = ref<Partial<Artist>>()
-
-	const imageFile = ref<File | null>(null)
 	const imagePreview = ref<string | null>(null)
-	const isDragging = ref(false)
-	const fileInput = ref<HTMLInputElement | null>(null)
-	const birthdayInputDate = useTemplateRef('birthdayInputDate')
-	const debutInputDate = useTemplateRef('debutInputDate')
 
-	const heroImageSrc = computed(() => {
-		return imagePreview.value || artistToEdit.value?.image || artist.value?.image || null
+	const effectiveHeroImageSrc = computed(() => {
+		return imagePreview.value || heroImageSrc.value
 	})
 
-	const overviewBadges = computed(() => {
-		if (!artistToEdit.value) return []
-
-		return [
-			{
-				label: artistTypeLabels[artistToEdit.value.type ?? 'SOLO'],
-				class: 'bg-cb-primary-900/15 text-cb-primary-300 ring-cb-primary-900/30',
-			},
-			{
-				label: genderLabels[artistToEdit.value.gender ?? 'UNKNOWN'],
-				class: 'bg-cb-quinary-900 text-white ring-cb-quinary-800',
-			},
-			{
-				label: artistToEdit.value.active_career ? 'Active career' : 'Inactive career',
-				class: artistToEdit.value.active_career
-					? 'bg-emerald-500/15 text-emerald-300 ring-emerald-500/30'
-					: 'bg-zinc-700/60 text-zinc-200 ring-zinc-600',
-			},
-		]
-	})
-
-	const overviewTaxonomyBadges = computed(() => {
-		return [
-			...artistNationalities.value.map((nationality) => ({
-				label: nationality.name,
-				class: 'bg-amber-500/15 text-amber-200 ring-amber-500/30',
-			})),
-			...artistStyles.value.map((style) => ({
-				label: style.name,
-				class: 'bg-cb-quinary-900 text-white ring-cb-quinary-800',
-			})),
-		]
-	})
-
-	const overviewStats = computed(() => {
-		if (!artistToEdit.value) return []
-
-		return [
-			{
-				label: artistToEdit.value.type === 'GROUP' ? 'Members' : 'Groups',
-				value: String(
-					artistToEdit.value.type === 'GROUP'
-						? artistMembers.value.length
-						: artistGroups.value.length,
-				),
-				helper: artistToEdit.value.type === 'GROUP' ? 'artist relations' : 'group links',
-			},
-			{
-				label: 'Companies',
-				value: String(artistCompanies.value.length),
-				helper: `${artistCompanies.value.filter((relation) => relation.is_current).length} current`,
-			},
-			{
-				label: 'Links',
-				value: String(artistPlatformList.value.length + artistSocialList.value.length),
-				helper: 'platforms and socials',
-			},
-		]
-	})
-
-	function onFileChange(e: Event) {
-		const files = (e.target as HTMLInputElement).files
-		if (files && files[0]) {
-			imageFile.value = files[0]
-			const reader = new FileReader()
-			reader.onload = (ev) => {
-				imagePreview.value = ev.target?.result as string
-			}
-			reader.readAsDataURL(files[0])
+	const onFileChange = (file: File) => {
+		const reader = new FileReader()
+		reader.onload = (ev) => {
+			imagePreview.value = ev.target?.result as string
 		}
+		reader.readAsDataURL(file)
 	}
 
-	function onDrop(e: DragEvent) {
-		isDragging.value = false
-		if (e.dataTransfer?.files?.length) {
-			const file = e.dataTransfer.files[0]
-			if (file) {
-				imageFile.value = file
-				const reader = new FileReader()
-				reader.onload = (ev) => {
-					imagePreview.value = ev.target?.result as string
-				}
-				reader.readAsDataURL(file)
-			}
-		}
+	const onFileDrop = (file: File) => {
+		onFileChange(file)
 	}
 
-	const sendUpdateArtist = async () => {
-		isUploadingEdit.value = true
-
-		try {
-			// Validation simple
-			if (!artistToEdit.value?.name) {
-				toast.add({ title: 'Name is required', color: 'error' })
-				isUploadingEdit.value = false
-				return
-			}
-
-			const currentArtistId = artist.value?.id
-			if (!currentArtistId) {
-				toast.add({ title: 'Artist not loaded', color: 'error' })
-				isUploadingEdit.value = false
-				return
-			}
-
-			const updates: Partial<Artist> = {
-				name: artistToEdit.value?.name,
-				// image: artistToEdit.value?.image, // the image n'est not modifiable ici
-				description: artistToEdit.value?.description,
-				id_youtube_music: artistToEdit.value?.id_youtube_music,
-				type: artistToEdit.value?.type,
-				gender: artistToEdit.value?.gender,
-				active_career: artistToEdit.value?.active_career,
-				// Re-convert CalendarDate to ISO string for the database
-				birth_date: birthdayToDate.value
-					? new Date(birthdayToDate.value.toString()).toISOString()
-					: null,
-				debut_date: debutDateToDate.value
-					? new Date(debutDateToDate.value.toString()).toISOString()
-					: null,
-				// Map selected objects to get names (or IDs if backend changed)
-				styles: artistStyles.value.map((style) => style.name),
-				general_tags: artistTags.value.map((tag) => tag.name),
-				nationalities: artistNationalities.value.map((nationality) => nationality.name),
-			}
-
-			const selectedCompanies = buildCompanyRelationsPayload()
-
-			// Filter out empty platform and social links before sending
-			const validPlatformLinks = platformLinkManager.getValidLinks()
-			const validSocialLinks = socialLinkManager.getValidLinks()
-
-			// Keep an explicit timeout at page level too: if the full edit flow
-			// never resolves, the user gets feedback and the save button unlocks again.
-			await runMutation(
-				updateArtist(
-					currentArtistId,
-					updates,
-					validSocialLinks,
-					validPlatformLinks,
-					buildArtistRefs(artistGroups.value),
-					buildArtistRefs(artistMembers.value),
-					selectedCompanies,
-				),
-				'Artist update timed out. Please try again.',
-			)
-
-			toast.add({
-				title: 'Artist updated successfully',
-				color: 'success',
-			})
-
-			await router.push(`/artist/${route.params.id}`)
-		} catch (error: unknown) {
-			const errorMessage = error instanceof Error ? error.message : 'Unknown error'
-			toast.add({
-				title: 'Error updating artist',
-				description: errorMessage,
-				color: 'error',
-			})
-			console.error('Error updating artist', error)
-		} finally {
-			isUploadingEdit.value = false
-		}
-	}
-	const { adjustTextareaDirect } = useTextareaAutoResize()
-
-	const handleNationalitiesUpdated = async () => {
-		nationalitiesList.value = await getAllNationalities()
-	}
-
-	const bootstrapEditor = async () => {
-		isBootstrapping.value = true
-		bootstrapError.value = null
-
-		try {
-			// The first editor screen depends on multiple async sources.
-			// Wrapping the whole Promise.all keeps the hero loader from hanging forever.
-			const [fullArtist, styles, tags, nationalities] = await runMutation(
-				Promise.all([
-					getFullArtistById(route.params.id as string),
-					getAllMusicStyles(),
-					getAllGeneralTags(),
-					getAllNationalities(),
-				]),
-				'Artist editor loading timed out. Please try again.',
-			)
-
-			artist.value = fullArtist
-			applyOptions({ styles, tags, nationalities })
-
-			if (artist.value) {
-				artistToEdit.value = { ...artist.value }
-
-				try {
-					// Links are loaded separately from the main artist payload, so they get
-					// their own timeout and their own warning state.
-					const { socialLinks, platformLinks } = await runMutation(
-						getSocialAndPlatformLinksByArtistId(artist.value.id),
-						'Artist links loading timed out. Please try again.',
-					)
-
-					// S'assurer que the tableaux ne sont jamais undefined
-					platformLinkManager.reset(
-						platformLinks && platformLinks.length > 0 ? platformLinks : [],
-					)
-					socialLinkManager.reset(
-						socialLinks && socialLinks.length > 0 ? socialLinks : [],
-					)
-				} catch (linkError) {
-					console.error('Error loading social and platform links:', linkError)
-					// Initialize with empty arrays if an error occurs
-					platformLinkManager.reset([])
-					socialLinkManager.reset([])
-					toast.add({
-						title: 'Warning',
-						description: 'Could not load all artist links',
-						color: 'warning',
-					})
-				}
-				applyArtistSelections(artist.value)
-
-				// Initialize CalendarDate values from the artist dates
-				birthdayToDate.value = artist.value.birth_date
-					? new Date(artist.value.birth_date)
-					: null
-				debutDateToDate.value = artist.value.debut_date
-					? new Date(artist.value.debut_date)
-					: null
-
-				if (import.meta.client) {
-					const textarea = document.querySelector('textarea')
-					if (textarea) {
-						adjustTextareaDirect(textarea)
-					}
-				}
-
-				title.value = 'EDIT ARTIST : ' + artist.value.name
-				description.value = artist.value.description || ''
-			}
-		} catch (error: unknown) {
-			const errorMessage = error instanceof Error ? error.message : 'Unknown error'
-			console.error('Error loading artist:', error)
-			bootstrapError.value = errorMessage
-			toast.add({
-				title: 'Error loading artist',
-				description: errorMessage,
-				color: 'error',
-			})
-		} finally {
-			isBootstrapping.value = false
-		}
+	const onNationalityCreated = async () => {
+		await refreshNationalities()
 	}
 
 	onMounted(async () => {
-		await bootstrapEditor()
+		await bootstrap()
 	})
 
 	useHead({
@@ -374,873 +63,21 @@
 </script>
 
 <template>
-	<PageHeroLoader
-		v-if="isBootstrapping"
-		variant="page"
-		title="Loading artist editor"
-		description="We are preparing the artist profile and taxonomy. Companies and relations will be searched on demand as you type."
+	<ArtistEditorShell
+		v-if="model"
+		mode="edit"
+		v-model="model"
+		:original="original"
+		:is-loading="isBootstrapping"
+		:bootstrap-error="bootstrapError"
+		:is-saving="isSaving"
+		:can-save="canSave"
+		:hero-image-src="effectiveHeroImageSrc"
+		:show-image-upload="isAdminStore"
+		@save="save"
+		@retry="bootstrap"
+		@image-change="onFileChange"
+		@image-drop="onFileDrop"
+		@nationality-created="onNationalityCreated"
 	/>
-
-	<div
-		v-else-if="bootstrapError"
-		class="flex min-h-[calc(100vh-140px)] items-center justify-center px-4 py-6"
-	>
-		<div
-			class="bg-cb-secondary-950 border-cb-quinary-900/70 w-full max-w-2xl rounded-[28px] border p-10 shadow-2xl"
-		>
-			<div class="flex flex-col items-center gap-5 text-center">
-				<div
-					class="bg-cb-primary-900/15 text-cb-primary-300 ring-cb-primary-900/30 rounded-2xl px-4 py-2 text-sm font-medium ring-1"
-				>
-					Editor loading failed
-				</div>
-				<div class="space-y-2">
-					<h1 class="text-2xl font-semibold">Artist editor is not ready yet</h1>
-					<p class="mx-auto max-w-xl text-sm leading-6 text-gray-400">
-						{{ bootstrapError }}
-					</p>
-				</div>
-				<UButton
-					label="Retry loading"
-					icon="i-lucide-refresh-cw"
-					color="primary"
-					class="!bg-cb-primary-900 hover:!bg-cb-primary-800 cursor-pointer justify-center !text-white hover:!text-white"
-					@click="bootstrapEditor"
-				/>
-			</div>
-		</div>
-	</div>
-
-	<div
-		v-else-if="artistToEdit && artist"
-		class="mx-auto min-h-[calc(100vh-60px)] max-w-7xl space-y-6 px-4 py-6 sm:px-6 lg:px-8"
-	>
-		<section
-			class="bg-cb-secondary-950 border-cb-quinary-900/70 overflow-hidden rounded-[28px] border shadow-2xl"
-		>
-			<div
-				class="border-cb-quinary-900/70 flex flex-col gap-6 border-b px-6 py-6 xl:flex-row xl:items-start xl:justify-between"
-			>
-				<div class="flex flex-col gap-5 sm:flex-row sm:items-start">
-					<div
-						class="bg-cb-quinary-900 border-cb-quinary-900/70 flex h-28 w-28 shrink-0 items-center justify-center overflow-hidden rounded-3xl border"
-					>
-						<NuxtImg
-							v-if="heroImageSrc"
-							:src="heroImageSrc"
-							:alt="artistToEdit.name"
-							format="webp"
-							loading="lazy"
-							class="h-full w-full object-cover"
-						/>
-						<UIcon v-else name="i-lucide-image" class="text-cb-quinary-700 h-10 w-10" />
-					</div>
-
-					<div class="space-y-4">
-						<div class="space-y-2">
-							<p
-								class="text-cb-quinary-700 text-xs font-semibold tracking-[0.35em] uppercase"
-							>
-								Artist editor
-							</p>
-							<div class="space-y-1">
-								<h1 class="text-2xl font-bold sm:text-3xl">
-									{{ artistToEdit.name || 'Untitled artist' }}
-								</h1>
-								<p class="max-w-2xl text-sm leading-6 text-gray-400">
-									Refine core identity, relationships, companies and editorial data from a
-									single workspace.
-								</p>
-							</div>
-						</div>
-
-						<div class="flex flex-wrap gap-2">
-							<span
-								v-for="badge in overviewBadges"
-								:key="badge.label"
-								:class="badge.class"
-								class="rounded-full px-3 py-1 text-xs font-medium ring-1"
-							>
-								{{ badge.label }}
-							</span>
-						</div>
-
-						<div v-if="overviewTaxonomyBadges.length > 0" class="flex flex-wrap gap-2">
-							<span
-								v-for="badge in overviewTaxonomyBadges"
-								:key="badge.label"
-								:class="badge.class"
-								class="rounded-full px-3 py-1 text-xs font-medium ring-1"
-							>
-								{{ badge.label }}
-							</span>
-						</div>
-
-						<div class="flex flex-wrap gap-2 text-sm text-gray-300">
-							<div
-								class="bg-cb-quaternary-950 border-cb-quinary-900/70 rounded-full border px-3 py-1.5"
-							>
-								<span class="text-cb-quinary-700 mr-2 text-xs tracking-[0.2em] uppercase">
-									Artist ID
-								</span>
-								<span class="font-medium">{{ artistToEdit.id }}</span>
-							</div>
-							<div
-								class="bg-cb-quaternary-950 border-cb-quinary-900/70 rounded-full border px-3 py-1.5"
-							>
-								<span class="text-cb-quinary-700 mr-2 text-xs tracking-[0.2em] uppercase">
-									YouTube
-								</span>
-								<span class="font-medium">
-									{{ artistToEdit.id_youtube_music || 'Not linked yet' }}
-								</span>
-							</div>
-						</div>
-					</div>
-				</div>
-
-				<div class="flex w-full flex-col gap-3 xl:w-auto xl:min-w-[240px]">
-					<UButton
-						label="View artist page"
-						icon="i-lucide-eye"
-						color="neutral"
-						variant="soft"
-						class="w-full cursor-pointer justify-center"
-						:to="`/artist/${artist.id}`"
-					/>
-					<UButton
-						label="Save changes"
-						icon="i-lucide-save"
-						color="primary"
-						:loading="isUploadingEdit"
-						class="!bg-cb-primary-900 hover:!bg-cb-primary-800 disabled:!bg-cb-primary-900 w-full cursor-pointer justify-center !text-white hover:!text-white disabled:!text-white"
-						@click="sendUpdateArtist"
-					/>
-					<p class="text-xs leading-5 text-gray-500">
-						Changes are applied directly to the artist record and related junction tables.
-					</p>
-				</div>
-			</div>
-
-			<ArtistOverviewStats :stats="overviewStats" />
-		</section>
-
-		<div class="grid gap-6 xl:grid-cols-[minmax(0,1.6fr)_minmax(320px,0.95fr)]">
-			<div class="space-y-6">
-				<section
-					class="bg-cb-secondary-950 border-cb-quinary-900/70 rounded-[28px] border p-6 shadow-xl"
-				>
-					<div class="mb-5 space-y-2">
-						<h2 class="text-xl font-semibold">Identity details</h2>
-						<p class="text-sm leading-6 text-gray-400">
-							Core identifiers used across pages, search and YouTube sync flows.
-						</p>
-					</div>
-
-					<div class="grid gap-4 lg:grid-cols-2">
-						<ComebackInput
-							v-model="artistToEdit.id"
-							label="Unique ID"
-							:placeholder="artist.id"
-							disabled
-						/>
-						<ComebackInput
-							v-model="artistToEdit.name"
-							label="Display name"
-							:placeholder="artist.name"
-						/>
-						<div class="lg:col-span-2">
-							<ComebackInput
-								:model-value="artistToEdit.id_youtube_music || ''"
-								label="YouTube Music ID"
-								:placeholder="artist.id_youtube_music || ''"
-								@update:model-value="
-									(value: string | number) => {
-										if (artistToEdit) {
-											artistToEdit.id_youtube_music = value ? String(value) : null
-										}
-									}
-								"
-							/>
-						</div>
-					</div>
-				</section>
-
-				<section
-					class="bg-cb-secondary-950 border-cb-quinary-900/70 rounded-[28px] border p-6 shadow-xl"
-				>
-					<div class="mb-5 space-y-2">
-						<h2 class="text-xl font-semibold">Timeline and status</h2>
-						<p class="text-sm leading-6 text-gray-400">
-							Define how the artist should be classified and when the project became
-							active.
-						</p>
-					</div>
-
-					<div
-						class="grid gap-4"
-						:class="artistToEdit.type === 'GROUP' ? 'lg:grid-cols-1' : 'lg:grid-cols-2'"
-					>
-						<div v-if="artistToEdit.type !== 'GROUP'" class="space-y-2">
-							<ComebackLabel label="Birthday" />
-							<UInputDate
-								ref="birthdayInputDate"
-								v-model="birthdayInputValue"
-								:min-value="new CalendarDate(1900, 1, 1)"
-								locale="en-GB"
-								class="w-full"
-								:ui="{
-									base: 'bg-cb-quaternary-950 border border-cb-quinary-900/70 rounded-xl',
-								}"
-							>
-								<template #trailing>
-									<UPopover
-										:reference="
-											birthdayInputDate?.inputsRef?.[3]?.$el ??
-											birthdayInputDate?.inputsRef?.[2]?.$el
-										"
-									>
-										<UButton
-											color="neutral"
-											variant="link"
-											size="sm"
-											icon="i-lucide-calendar"
-											aria-label="Select a birthday date"
-											class="cursor-pointer px-0"
-										/>
-										<template #content>
-											<UCalendar
-												v-model="birthdayInputValue"
-												class="bg-cb-quinary-900 rounded p-2"
-												:min-value="new CalendarDate(1900, 1, 1)"
-												locale="en-GB"
-												year-controls
-											/>
-										</template>
-									</UPopover>
-								</template>
-							</UInputDate>
-						</div>
-
-						<div class="space-y-2">
-							<ComebackLabel label="Debut date" />
-							<UInputDate
-								ref="debutInputDate"
-								v-model="debutDateInputValue"
-								:min-value="new CalendarDate(2000, 1, 1)"
-								locale="en-GB"
-								class="w-full"
-								:ui="{
-									base: 'bg-cb-quaternary-950 border border-cb-quinary-900/70 rounded-xl',
-								}"
-							>
-								<template #trailing>
-									<UPopover
-										:reference="
-											debutInputDate?.inputsRef?.[3]?.$el ??
-											debutInputDate?.inputsRef?.[2]?.$el
-										"
-									>
-										<UButton
-											color="neutral"
-											variant="link"
-											size="sm"
-											icon="i-lucide-calendar"
-											aria-label="Select a debut date"
-											class="cursor-pointer px-0"
-										/>
-										<template #content>
-											<UCalendar
-												v-model="debutDateInputValue"
-												class="bg-cb-quinary-900 rounded p-2"
-												:min-value="new CalendarDate(2000, 1, 1)"
-												locale="en-GB"
-												year-controls
-											/>
-										</template>
-									</UPopover>
-								</template>
-							</UInputDate>
-						</div>
-					</div>
-
-					<div class="mt-6 grid gap-5 xl:grid-cols-3">
-						<div class="space-y-3">
-							<ComebackLabel label="Gender" />
-							<div class="flex flex-wrap gap-2">
-								<UButton
-									v-for="option in genderOptions"
-									:key="option.value"
-									type="button"
-									size="sm"
-									:color="artistToEdit.gender === option.value ? 'primary' : 'neutral'"
-									:variant="artistToEdit.gender === option.value ? 'solid' : 'soft'"
-									class="cursor-pointer rounded-full"
-									:aria-pressed="artistToEdit.gender === option.value"
-									@click="artistToEdit.gender = option.value"
-								>
-									{{ option.label }}
-								</UButton>
-							</div>
-						</div>
-
-						<div class="space-y-3">
-							<ComebackLabel label="Type" />
-							<div class="flex flex-wrap gap-2">
-								<UButton
-									v-for="option in artistTypeOptions"
-									:key="option.value"
-									type="button"
-									size="sm"
-									:color="artistToEdit.type === option.value ? 'primary' : 'neutral'"
-									:variant="artistToEdit.type === option.value ? 'solid' : 'soft'"
-									class="cursor-pointer rounded-full"
-									:aria-pressed="artistToEdit.type === option.value"
-									@click="artistToEdit.type = option.value"
-								>
-									{{ option.label }}
-								</UButton>
-							</div>
-						</div>
-
-						<div class="space-y-3">
-							<ComebackLabel label="Career status" />
-							<div class="flex flex-wrap gap-2">
-								<UButton
-									v-for="option in careerOptions"
-									:key="option.label"
-									type="button"
-									size="sm"
-									:color="
-										artistToEdit.active_career === option.value ? 'primary' : 'neutral'
-									"
-									:variant="
-										artistToEdit.active_career === option.value ? 'solid' : 'soft'
-									"
-									class="cursor-pointer rounded-full"
-									:aria-pressed="artistToEdit.active_career === option.value"
-									@click="artistToEdit.active_career = option.value"
-								>
-									{{ option.label }}
-								</UButton>
-							</div>
-						</div>
-					</div>
-				</section>
-
-				<section
-					class="bg-cb-secondary-950 border-cb-quinary-900/70 rounded-[28px] border p-6 shadow-xl"
-				>
-					<div class="mb-5 space-y-2">
-						<h2 class="text-xl font-semibold">Taxonomy</h2>
-						<p class="text-sm leading-6 text-gray-400">
-							Use styles, nationalities and tags to improve filtering, discovery and
-							cross-linking.
-						</p>
-					</div>
-
-					<div class="grid gap-5 xl:grid-cols-2">
-						<div v-if="stylesList" class="space-y-3 xl:col-span-full">
-							<div class="flex flex-wrap items-center justify-between gap-3">
-								<ComebackLabel label="Styles" />
-								<UModal
-									:ui="{
-										overlay: 'bg-cb-quinary-950/75',
-										content: 'ring-cb-quinary-950',
-									}"
-								>
-									<UButton
-										label="Create new style"
-										variant="soft"
-										color="primary"
-										class="cursor-pointer"
-									/>
-
-									<template #content>
-										<ModalCreateStyleTag :style-fetch="stylesList" />
-									</template>
-								</UModal>
-							</div>
-							<UInputMenu
-								v-model="artistStyles"
-								:items="stylesForMenu"
-								by="id"
-								multiple
-								placeholder="Select styles"
-								searchable
-								searchable-placeholder="Search a style..."
-								class="w-full"
-								:ui="{
-									base: 'bg-cb-quaternary-950 border border-cb-quinary-900/70 rounded-xl',
-									content: 'bg-cb-quaternary-950',
-									item: 'rounded cursor-pointer data-highlighted:before:bg-cb-primary-900/30 hover:bg-cb-primary-900',
-								}"
-							/>
-						</div>
-
-						<div v-if="nationalitiesList" class="space-y-3">
-							<div class="flex flex-wrap items-center justify-between gap-3">
-								<ComebackLabel label="Nationalities" />
-								<UModal
-									:ui="{
-										overlay: 'bg-cb-quinary-950/75',
-										content: 'ring-cb-quinary-950',
-									}"
-								>
-									<UButton
-										label="Create new nationality"
-										variant="soft"
-										color="primary"
-										class="cursor-pointer"
-									/>
-
-									<template #content>
-										<ModalCreateNationality
-											:nationalities="nationalitiesList"
-											@created="handleNationalitiesUpdated"
-										/>
-									</template>
-								</UModal>
-							</div>
-							<UInputMenu
-								v-model="artistNationalities"
-								:items="nationalitiesForMenu"
-								by="id"
-								multiple
-								placeholder="Select nationalities"
-								searchable
-								searchable-placeholder="Search a nationality..."
-								class="w-full"
-								:ui="{
-									base: 'bg-cb-quaternary-950 border border-cb-quinary-900/70 rounded-xl',
-									content: 'bg-cb-quaternary-950',
-									item: 'rounded cursor-pointer data-highlighted:before:bg-cb-primary-900/30 hover:bg-cb-primary-900',
-								}"
-							/>
-						</div>
-
-						<div v-if="tagsList" class="space-y-3">
-							<div class="flex flex-wrap items-center justify-between gap-3">
-								<ComebackLabel label="General tags" />
-								<UModal
-									:ui="{
-										overlay: 'bg-cb-quinary-950/75',
-										content: 'ring-cb-quinary-950',
-									}"
-								>
-									<UButton
-										label="Create new tag"
-										variant="soft"
-										color="primary"
-										class="cursor-pointer"
-									/>
-
-									<template #content>
-										<ModalCreateTag :general-tags="tagsList" />
-									</template>
-								</UModal>
-							</div>
-							<UInputMenu
-								v-model="artistTags"
-								:items="tagsForMenu"
-								by="id"
-								multiple
-								placeholder="Select tags"
-								searchable
-								searchable-placeholder="Search a tag..."
-								class="w-full"
-								:ui="{
-									base: 'bg-cb-quaternary-950 border border-cb-quinary-900/70 rounded-xl',
-									content: 'bg-cb-quaternary-950',
-									item: 'rounded cursor-pointer data-highlighted:before:bg-cb-primary-900/30 hover:bg-cb-primary-900',
-								}"
-							/>
-						</div>
-					</div>
-				</section>
-
-				<section
-					class="bg-cb-secondary-950 border-cb-quinary-900/70 rounded-[28px] border p-6 shadow-xl"
-				>
-					<div class="mb-5 space-y-2">
-						<h2 class="text-xl font-semibold">Editorial description</h2>
-						<p class="text-sm leading-6 text-gray-400">
-							Keep the page copy concise and readable. This text is used on the public
-							artist page and in search contexts.
-						</p>
-					</div>
-
-					<textarea
-						v-model="artistToEdit.description"
-						:placeholder="
-							artistToEdit.description || 'Write a concise artist description...'
-						"
-						class="bg-cb-quaternary-950 border-cb-quinary-900/70 focus:border-cb-primary-900/60 focus:ring-cb-primary-900/20 min-h-[220px] w-full rounded-2xl border p-4 text-sm leading-6 text-white transition outline-none focus:ring-2"
-						@input="adjustTextareaDirect($event.target as HTMLTextAreaElement)"
-					/>
-				</section>
-
-				<section
-					class="bg-cb-secondary-950 border-cb-quinary-900/70 rounded-[28px] border p-6 shadow-xl"
-				>
-					<div class="mb-5 flex flex-wrap items-start justify-between gap-3">
-						<div class="space-y-2">
-							<h2 class="text-xl font-semibold">Artist relationships</h2>
-							<p class="text-sm leading-6 text-gray-400">
-								Link soloists to their groups or manage the full roster for group
-								profiles.
-							</p>
-						</div>
-						<UModal
-							v-if="isAdminStore"
-							:ui="{
-								overlay: 'bg-cb-quinary-950/75',
-								content: 'ring-cb-quinary-950',
-							}"
-						>
-							<UButton
-								label="Create new artist"
-								variant="soft"
-								color="primary"
-								class="cursor-pointer"
-							/>
-
-							<template #content>
-								<ModalCreateArtist
-									:styles-list="stylesList"
-									:nationalities-list="nationalitiesList"
-									:tags-list="tagsList"
-								/>
-							</template>
-						</UModal>
-					</div>
-
-					<div
-						class="grid gap-5"
-						:class="artistToEdit.type === 'GROUP' ? 'xl:grid-cols-2' : ''"
-					>
-						<div class="space-y-3">
-							<ComebackLabel
-								:label="artistToEdit.type === 'GROUP' ? 'Related groups' : 'Groups'"
-							/>
-							<UInputMenu
-								v-model="artistGroups"
-								:search-term="groupSearchTerm"
-								:items="groupsForMenu"
-								by="id"
-								multiple
-								placeholder="Search groups"
-								searchable
-								searchable-placeholder="Search a group..."
-								:loading="isSearchingGroups"
-								class="w-full"
-								:ui="{
-									base: 'bg-cb-quaternary-950 border border-cb-quinary-900/70 rounded-xl',
-									content: 'bg-cb-quaternary-950',
-									item: 'rounded cursor-pointer data-highlighted:before:bg-cb-primary-900/30 hover:bg-cb-primary-900',
-								}"
-								@update:search-term="onGroupSearchTermChange"
-							/>
-						</div>
-
-						<div v-if="artistToEdit.type === 'GROUP'" class="space-y-3">
-							<ComebackLabel label="Members" />
-							<UInputMenu
-								v-model="artistMembers"
-								:search-term="memberSearchTerm"
-								:items="membersForMenu"
-								by="id"
-								multiple
-								placeholder="Search members"
-								searchable
-								searchable-placeholder="Search a member..."
-								:loading="isSearchingMembers"
-								class="w-full"
-								:ui="{
-									base: 'bg-cb-quaternary-950 border border-cb-quinary-900/70 rounded-xl',
-									content: 'bg-cb-quaternary-950',
-									item: 'rounded cursor-pointer data-highlighted:before:bg-cb-primary-900/30 hover:bg-cb-primary-900',
-								}"
-								@update:search-term="onMemberSearchTermChange"
-							/>
-						</div>
-					</div>
-				</section>
-
-				<section
-					class="bg-cb-secondary-950 border-cb-quinary-900/70 rounded-[28px] border p-6 shadow-xl"
-				>
-					<div class="mb-5 flex flex-wrap items-start justify-between gap-3">
-						<div class="space-y-2">
-							<h2 class="text-xl font-semibold">Company relations</h2>
-							<p class="text-sm leading-6 text-gray-400">
-								Track labels, agencies and other business links tied to the artist
-								profile.
-							</p>
-						</div>
-						<div class="flex flex-wrap gap-2">
-							<UModal
-								v-model:open="isCompanyModalOpen"
-								:ui="{
-									overlay: 'bg-cb-quinary-950/75',
-									content: 'ring-cb-quinary-950',
-								}"
-							>
-								<UButton
-									label="Create new company"
-									variant="soft"
-									color="primary"
-									class="cursor-pointer"
-								/>
-
-								<template #content>
-									<ModalCreateEditCompany
-										:company="null"
-										:is-creating="true"
-										@updated="handleCompanyUpdated"
-									/>
-								</template>
-							</UModal>
-							<UButton
-								label="Add relation"
-								icon="i-lucide-plus"
-								color="primary"
-								class="cursor-pointer"
-								@click="addCompanyRelation"
-							/>
-						</div>
-					</div>
-
-					<div
-						v-if="artistCompanies.length > 0"
-						class="grid grid-cols-1 gap-4 xl:grid-cols-2"
-					>
-						<div
-							v-for="(relation, index) in artistCompanies"
-							:key="index"
-							class="bg-cb-quaternary-950 border-cb-quinary-900/70 rounded-2xl border p-4"
-						>
-							<div class="flex items-start justify-between gap-4">
-								<div class="min-w-0 flex-1 space-y-4">
-									<div class="space-y-2">
-										<label
-											class="text-cb-quinary-700 block text-xs font-semibold tracking-[0.2em] uppercase"
-										>
-											Company
-										</label>
-										<UInputMenu
-											:model-value="relation.company ?? undefined"
-											:search-term="companySearchTerm"
-											:items="companiesForMenu"
-											by="id"
-											placeholder="Select a company"
-											searchable
-											searchable-placeholder="Search company..."
-											:loading="isSearchingCompanies"
-											class="w-full"
-											:ui="{
-												base: 'bg-cb-secondary-950 border border-cb-quinary-900/70 rounded-xl',
-												content: 'bg-cb-secondary-950',
-												item: 'rounded cursor-pointer data-highlighted:before:bg-cb-primary-900/30 hover:bg-cb-primary-900',
-											}"
-											@update:model-value="
-												(company: unknown) =>
-													updateCompanyInRelation(index, company as CompanyMenuItem)
-											"
-											@update:search-term="onCompanySearchTermChange"
-										/>
-									</div>
-
-									<div class="space-y-2">
-										<label
-											class="text-cb-quinary-700 block text-xs font-semibold tracking-[0.2em] uppercase"
-										>
-											Relationship type
-										</label>
-										<select
-											v-model="relation.relationship_type"
-											class="bg-cb-secondary-950 border-cb-quinary-900/70 focus:border-cb-primary-900/60 focus:ring-cb-primary-900/20 w-full rounded-xl border px-3 py-2.5 text-sm transition outline-none focus:ring-2"
-										>
-											<option v-for="type in relationshipTypes" :key="type" :value="type">
-												{{ type }}
-											</option>
-										</select>
-									</div>
-
-									<label
-										:for="`current-${index}`"
-										class="border-cb-quinary-900/70 flex cursor-pointer items-center gap-3 rounded-xl border px-3 py-2 text-sm text-gray-300"
-									>
-										<input
-											:id="`current-${index}`"
-											v-model="relation.is_current"
-											type="checkbox"
-											class="accent-cb-primary-900 h-4 w-4 rounded"
-										/>
-										<span>Current relationship</span>
-									</label>
-								</div>
-
-								<UButton
-									icon="i-lucide-trash-2"
-									color="error"
-									variant="soft"
-									class="cursor-pointer"
-									aria-label="Remove company relation"
-									title="Remove company relation"
-									@click="removeCompanyRelation(index)"
-								/>
-							</div>
-						</div>
-					</div>
-
-					<div
-						v-else
-						class="bg-cb-quaternary-950 border-cb-quinary-900/70 rounded-2xl border border-dashed px-4 py-10 text-center text-sm text-gray-400"
-					>
-						No company relations yet. Add one when the artist is tied to a label, agency
-						or distributor.
-					</div>
-				</section>
-
-				<section
-					class="bg-cb-secondary-950 border-cb-quinary-900/70 rounded-[28px] border p-6 shadow-xl"
-				>
-					<div class="mb-5 space-y-2">
-						<h2 class="text-xl font-semibold">Platforms and socials</h2>
-						<p class="text-sm leading-6 text-gray-400">
-							Keep official links updated so cards and profile pages stay accurate.
-						</p>
-					</div>
-
-					<div class="grid grid-cols-1 gap-5 lg:grid-cols-2">
-						<LinkManager
-							:items="artistPlatformList"
-							label="Platforms"
-							name-placeholder="Platform's name"
-							link-placeholder="Platform's link"
-							key-prefix="platform"
-							@add-item="platformLinkManager.add"
-							@remove-item="platformLinkManager.remove"
-							@update-name="platformLinkManager.updateName"
-							@update-link="platformLinkManager.updateLink"
-						/>
-
-						<LinkManager
-							:items="artistSocialList"
-							label="Socials"
-							name-placeholder="Social name"
-							link-placeholder="Social link"
-							key-prefix="social"
-							@add-item="socialLinkManager.add"
-							@remove-item="socialLinkManager.remove"
-							@update-name="socialLinkManager.updateName"
-							@update-link="socialLinkManager.updateLink"
-						/>
-					</div>
-				</section>
-			</div>
-
-			<div class="space-y-6 xl:sticky xl:top-24 xl:self-start">
-				<section
-					class="bg-cb-secondary-950 border-cb-quinary-900/70 rounded-[28px] border p-6 shadow-xl"
-				>
-					<div class="mb-4 space-y-2">
-						<h2 class="text-xl font-semibold">Visuals and sync</h2>
-						<p class="text-sm leading-6 text-gray-400">
-							The public profile image normally follows YouTube Music. Admins can stage a
-							custom preview here before saving.
-						</p>
-					</div>
-
-					<div
-						class="bg-cb-quaternary-950 border-cb-quinary-900/70 mb-4 overflow-hidden rounded-3xl border"
-					>
-						<NuxtImg
-							v-if="heroImageSrc"
-							:src="heroImageSrc"
-							:alt="artistToEdit.name"
-							format="webp"
-							loading="lazy"
-							class="aspect-[4/3] w-full object-cover"
-						/>
-						<div
-							v-else
-							class="text-cb-quinary-700 flex aspect-[4/3] items-center justify-center"
-						>
-							<UIcon name="i-lucide-image" class="h-12 w-12" />
-						</div>
-					</div>
-
-					<UFormField
-						v-if="isAdminStore"
-						label="Custom image preview"
-						description="Drop a file here to preview a custom image before saving."
-					>
-						<div
-							:class="{ 'bg-cb-primary-900/15 border-cb-primary-900/60': isDragging }"
-							class="bg-cb-quaternary-950 border-cb-quinary-900/70 cursor-pointer rounded-2xl border border-dashed p-5 text-center transition"
-							@click="fileInput?.click()"
-							@dragover.prevent="isDragging = true"
-							@dragleave.prevent="isDragging = false"
-							@drop.prevent="onDrop"
-						>
-							<input
-								ref="fileInput"
-								type="file"
-								accept="image/*"
-								class="hidden"
-								@change="onFileChange"
-							/>
-							<p class="text-sm text-gray-300">
-								Drag and drop an image here, or click to browse from disk.
-							</p>
-							<p class="mt-2 text-xs text-gray-500">
-								This preview does not bypass the regular YouTube sync behavior.
-							</p>
-						</div>
-					</UFormField>
-
-					<div
-						v-else
-						class="bg-cb-quaternary-950 border-cb-quinary-900/70 rounded-2xl border p-4 text-sm leading-6 text-gray-400"
-					>
-						Image updates are synchronized automatically from YouTube Music for non-admin
-						users.
-					</div>
-				</section>
-
-				<ArtistQuickOverview
-					description="Useful checkpoints before publishing your edits."
-					:nationalities-count="artistNationalities.length"
-					:profile-type="artistToEdit.type"
-					:birthday-label="formatDisplayDate(birthdayToDate)"
-					:debut-date-label="formatDisplayDate(debutDateToDate)"
-					:tags-count="artistTags.length"
-				/>
-
-				<ArtistSavePanel
-					description="Use this primary action once the profile feels consistent."
-				>
-					<UButton
-						label="Save changes"
-						icon="i-lucide-save"
-						color="primary"
-						size="xl"
-						:loading="isUploadingEdit"
-						class="!bg-cb-primary-900 hover:!bg-cb-primary-800 disabled:!bg-cb-primary-900 w-full cursor-pointer justify-center !text-white hover:!text-white disabled:!text-white"
-						@click="sendUpdateArtist"
-					/>
-					<UButton
-						label="Return to artist page"
-						icon="i-lucide-arrow-left"
-						color="neutral"
-						variant="soft"
-						class="w-full cursor-pointer justify-center"
-						:to="`/artist/${artist.id}`"
-					/>
-				</ArtistSavePanel>
-			</div>
-		</div>
-	</div>
 </template>
